+++ cmd/dive/go.mod
@@ -34,5 +34,6 @@ require (
 	github.com/tidwall/pretty v1.2.1 // indirect
 	github.com/tidwall/sjson v1.2.5 // indirect
 	github.com/yosida95/uritemplate/v3 v3.0.2 // indirect
+	go.jetify.com/typeid v1.3.0 // indirect
 	golang.org/x/sys v0.33.0 // indirect
 )
