+++ cmd/simple_runner/main.go
@@ -63,9 +63,9 @@ func main() {
 	}
 
 	// Choose the workflow to run
-	var workflow *workflow.Workflow
+	var w *workflow.Workflow
 	if workflowName != "" {
-		workflow, err = env.GetWorkflow(workflowName)
+		w, err = env.GetWorkflow(workflowName)
 		if err != nil {
 			fatal(err.Error())
 		}
@@ -74,13 +74,15 @@ func main() {
 		if len(workflows) != 1 {
 			fatal("You must specify a workflow name")
 		}
-		workflow = workflows[0]
+		w = workflows[0]
 	}
 
 	// Start the workflow
-	execution, err := env.ExecuteWorkflow(ctx, environment.ExecutionOptions{
-		WorkflowName: workflow.Name(),
-		Inputs:       vars,
+	execution, err := environment.NewEventBasedExecution(env, environment.ExecutionOptions{
+		WorkflowName:   w.Name(),
+		Inputs:         vars,
+		EventStore:     workflow.NewNullEventStore(),
+		EventBatchSize: 10,
 	})
 	if err != nil {
 		fatal(err.Error())
