+++ environment/dynamic_content.go
@@ -156,67 +156,6 @@ func (c *RisorContent) convertScriptResultToContent(result interface{}) ([]llm.C
 	}
 }
 
-// convertMapToContent converts a map to a Content object
-func (c *RisorContent) convertMapToContent(m map[string]interface{}) (llm.Content, error) {
-	contentType, ok := m["type"].(string)
-	if !ok {
-		return nil, fmt.Errorf("content block missing 'type' field")
-	}
-
-	switch contentType {
-	case "text":
-		text, ok := m["text"].(string)
-		if !ok {
-			return nil, fmt.Errorf("text content block missing 'text' field")
-		}
-		return &llm.TextContent{Text: text}, nil
-	case "image":
-		source, err := c.convertMapToContentSource(m)
-		if err != nil {
-			return nil, fmt.Errorf("failed to convert image source: %w", err)
-		}
-		return &llm.ImageContent{Source: source}, nil
-	case "document":
-		source, err := c.convertMapToContentSource(m)
-		if err != nil {
-			return nil, fmt.Errorf("failed to convert document source: %w", err)
-		}
-		return &llm.DocumentContent{Source: source}, nil
-	default:
-		return nil, fmt.Errorf("unsupported content type: %s", contentType)
-	}
-}
-
-// convertMapToContentSource converts a map to a ContentSource
-func (c *RisorContent) convertMapToContentSource(m map[string]interface{}) (*llm.ContentSource, error) {
-	source := &llm.ContentSource{}
-
-	if sourceType, ok := m["source_type"].(string); ok {
-		switch sourceType {
-		case "url":
-			source.Type = llm.ContentSourceTypeURL
-		case "base64":
-			source.Type = llm.ContentSourceTypeBase64
-		default:
-			return nil, fmt.Errorf("unsupported source type: %s", sourceType)
-		}
-	}
-
-	if mediaType, ok := m["media_type"].(string); ok {
-		source.MediaType = mediaType
-	}
-
-	if url, ok := m["url"].(string); ok {
-		source.URL = url
-	}
-
-	if data, ok := m["data"].(string); ok {
-		source.Data = data
-	}
-
-	return source, nil
-}
-
 //// ScriptPathContent /////////////////////////////////////////////////////////
 
 // ScriptPathContent represents content that is dynamically generated by executing
